pyproject.toml
[tool.poetry]
name = "pyconkor2021"
version = "0.1.0"
description = ""
authors = ["qu3vipon <qu3vipon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"
dj-rest-auth = "^2.1.7"
djangorestframework-simplejwt = "^4.7.1"
pytest = "^6.2.4"
pytest-django = "^4.4.0"
Schema = "^0.7.4"
pytest-lazy-fixture = "^0.6.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
